fix: extract metrics tracker
